requirements.txt
Flask
Flask-Bcrypt
PyJWT
mysql-connector-python
PyMySQL
cryptography
gunicorn